// src/app/models/competencia_model.ts
export class Competencia_Model {
	id: string;
    competencia: string;
    objetivo_mejora: string;
    reto01: string;
    reto02: string;
    crossTraining: boolean;
    crossTraining_texto: string;
    tutoria: boolean;
    tutoria_texto: string;
    formaciones: any;
    formacion_otro: boolean;
    formacion_otros: string;
}